pyproject.toml
[tool.poetry]
name = "social-network"
version = "0.1.0"
description = ""
authors = ["popkir <popkir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
psycopg2 = "^2.9.6"
sqlalchemy = "^2.0.12"
python-dotenv = "^1.0.0"
alembic = "^1.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
